Uploading the second commit
Added a new JTable component to show list of all employees and added layout for showing highlighted/selected employees from the table below. Not working right now so it won't populate the fields above. Having issue with JLabel icon causing the view employee panel to not load. Added the deleted button to the form but need to add the functions to it.