Next features
Added read capability. Will populate the employee view area with the row that has been selected from the table. There are issues like if all the fields are not filled with any text, in view product when a row is selected it will give the error for the missing text field. Also we can't add multiple values to the table so it overwrites the first row only. Have added delete so it will delete the selected row. Not added update search or icon. Icon seems to be having issues. Also in previous commit start date text field was missing. Confirmed with TA that we can use external JAR files but had issue with loading of the JDateChooser jar class. For now it is a text field area.